feat: detailsView
but LineChart is TODO :AJOUTE LE DERNIER PRIX DE L'ACTION SI IL N'Y A PAS DE TRANSACTION CE JOUR LA

+++ src/main/java/fr/cashcoders/capitalhub/controller/utils/DatabaseFeeder.java
@@ -62,7 +62,7 @@ public class DatabaseFeeder {
         ResultSet actionProduitResultSet = actionProduitStatement.executeQuery();
         while (actionProduitResultSet.next()) {
             int actionId = actionProduitResultSet.getInt("idaction");
-            int quantity = actionProduitResultSet.getInt("quantity");
+            double quantity = actionProduitResultSet.getDouble("quantity");
 
             PreparedStatement actionStatement = connection.prepareStatement("SELECT * FROM action WHERE id = ?");
             actionStatement.setInt(1, actionId);
